% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{A_to_triangle}
\alias{A_to_triangle}
\title{Apply symmetry from one trianlge region across other two regions.}
\usage{
A_to_triangle(triangle, A.species)
}
\value{
A list.
}
\description{
Applies symmetry from one trianlge region across other two regions.
Used by \code{\link{goelz}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
