% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{add_one}
\alias{add_one}
\title{Add one}
\usage{
add_one(x)
}
\value{
A numeric vector.
}
\description{
Adds one.
Used by \code{\link{goelz}}, \code{\link{goelz_add_border}}, and \code{\link{A_to_triangle}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
