% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{annotate_pop}
\alias{annotate_pop}
\title{Annotate population with fitness and stats}
\usage{
annotate_pop(population, fitness, gens, gen.ids)
}
\value{
A popluation matrix.
}
\description{
Annotates population with fitness and stats.
Used by \code{\link{goelz_optim}} & \code{\link{nelder_biculture_optim}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
