% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{assign_species}
\alias{assign_species}
\title{Assign speices identities in a Nelder Fan biculture design}
\usage{
assign_species(data, comps)
}
\arguments{
\item{data}{An object of class nelder or class nelder-biculture.}

\item{comps}{An optional numeric vector containing the ratios of one species in each spoke.}
}
\value{
An object of the same class as \code{data}.
}
\description{
Assigns speices identities in a Nelder Fan biculture design.
Used within \code{\link{nelder_biculture}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
