% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{compile_pop}
\alias{compile_pop}
\title{Compile popluation data}
\usage{
compile_pop(pop, fit, triangle, GEN)
}
\value{
A list.
}
\description{
Compiles popluation data. Used by \code{\link{goelz_optim}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
