% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{goelz}
\alias{goelz}
\title{Create a Goelz Triangle experimental design}
\usage{
goelz(N = 35)
}
\arguments{
\item{N}{The number of plants to be on each edge of the design.}
}
\value{
A data frame (tibble) of class sysd and class goelz.
}
\description{
Creates a Goelz Triangle experimental design.
}
\details{
The Goelz Triangle experimental design (Goelz 2001) is an experimental design that systematically
varies species composition of three plant species within a single plot, while maintaining a constant plant density.
This function creates a Goelz Triangle design following several criteria set forth by Goelz (2001):
\itemize{
 \item{"symmetry"}{ - Symmetry requires that, for every planting spot assigned, the 2 corresponding planting
 spots be assigned corresponding species.}
 \item{"equality"}{ - Equality merely requires that equal numbers of each species be assigned to each plot.
 This will be achieved if the number of planting spots per plot is a multiple of three and if symmetry is imposed.}
 \item{"conformity"}{ - Conformity to the intended pattern requires that the species proportion in any subsection
 of the triangular plot is close to the expectations.}
}
This function does NOT robustly impose the conformity criterion. Species are assigned randomly to each position
by sampling the three species using weights based on the theoretical probability of each species at that point.
For a robust implementation of the conformity criterion that optimizes the rough initial approach of this function
using an evolutionary algorithm, use \code{\link{goelz_optim}}.
}
\examples{
dat <- goelz()
}
\references{
\itemize{
 \item Goelz JCG (2001) Systematic experimental designs for mixed species plantings. Native Plants Journal 2:90–96.
 \url{http://npj.uwpress.org/content/2/2/90.short}
}
}
\seealso{
Other definition functions: 
\code{\link{goelz_add_border}()},
\code{\link{goelz_corners}()},
\code{\link{goelz_guides}()},
\code{\link{goelz_mirror}()},
\code{\link{goelz_optim}()},
\code{\link{goelz_starts}()},
\code{\link{nelder_biculture_competition}()},
\code{\link{nelder_biculture_optim}()},
\code{\link{nelder_biculture}()},
\code{\link{nelder_decision}()},
\code{\link{nelder_interspoke_distance}()},
\code{\link{nelder}()},
\code{\link{select_optimal_goelz}()},
\code{\link{select_optimal_nelder_biculture}()}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
