% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{goelz_add_border}
\alias{goelz_add_border}
\title{Add border rows to a Goelz Triangle experimental design}
\usage{
goelz_add_border(data, n)
}
\arguments{
\item{data}{An object of class goelz.}

\item{n}{The number of border rows to add on each side of the design.}
}
\value{
A data frame (tibble) of class sysd, goelz, and goelz-border.
}
\description{
Adds border rows to a Goelz Triangle experimental design.
}
\details{
Goelz (2001) suggests that all Goelz Triangle experimental designs be enclosed within a number of
border/buffer rows. These border rows reduce edge effects and thereby increase the number of plants with useable
data. The presence of border rows will also increases the number of individuals whose nearest neighbors are all
conspecifics. Goelz (2001) suggests that the species in each planting spot in the border rows be assigned as a
50:50 probability of the species in the two adjacent spots towards the interior of the triangle. This is the method
used here, with each additional border row being determined successively. The border rows are also held to the same
standard of symmetry across the three zones in the triangle.
}
\examples{
dat <- goelz()
dat.border <- goelz_add_border(data = dat, n = 3)
}
\references{
\itemize{
 \item Goelz JCG (2001) Systematic experimental designs for mixed species plantings. Native Plants Journal 2:90–96.
 \url{http://npj.uwpress.org/content/2/2/90.short}
}
}
\seealso{
Other definition functions: \code{\link{goelz_corners}},
  \code{\link{goelz_guides}}, \code{\link{goelz_mirror}},
  \code{\link{goelz_optim}}, \code{\link{goelz_starts}},
  \code{\link{goelz}},
  \code{\link{nelder_biculture_competition}},
  \code{\link{nelder_biculture_optim}},
  \code{\link{nelder_biculture}},
  \code{\link{nelder_decision}},
  \code{\link{nelder_interspoke_distance}},
  \code{\link{nelder}}, \code{\link{select_optimal_goelz}},
  \code{\link{select_optimal_nelder_biculture}}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
