% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{goelz_class_check}
\alias{goelz_class_check}
\title{Check if an object if of class goelz}
\usage{
goelz_class_check(data)
}
\value{
An error.
}
\description{
Checks if an object if of class goelz. Used by many goelz definition functions.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
