% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{goelz_count}
\alias{goelz_count}
\title{Count total plants in Goelz Triangle design.}
\usage{
goelz_count(N)
}
\value{
A data frame (tibble).
}
\description{
Counts total plants in Goelz Triangle design. Used by \code{\link{goelz}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
