% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{goelz_fitness}
\alias{goelz_fitness}
\title{Calculate fitness of Goelz Trianlge design}
\usage{
goelz_fitness(design, layout)
}
\value{
A numeric value.
}
\description{
Calculates fitness of Goelz Trianlge design.
Used by \code{\link{goelz_optim}} as the fitness function.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
