% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{goelz_optim}
\alias{goelz_optim}
\title{Create conformity-optimized Goelz Triangle experimental designs}
\usage{
goelz_optim(N = 35, MU = 100, LAMBDA = MU, MAX.GEN = 150,
  P.RECOMB = 1, RECOMB = 0.1, P.MUT = 1, MUT = 0.005)
}
\arguments{
\item{N}{The number of plants to be on each edge of the design.}

\item{MU}{The population size.}

\item{LAMBDA}{The number of offspring to produce in each generation.}

\item{MAX.GEN}{The number of generations to run the evolutionary algorithm.}

\item{P.RECOMB}{The probability of two parents to perform crossover.}

\item{RECOMB}{The crossover probability for each gene.}

\item{P.MUT}{The probability to apply mutation to a child.}

\item{MUT}{The mutation probability for each gene.}
}
\value{
An list containing:
\itemize{
 \item{"stats"}{ - A data frame (tibble) containing statistics on each generation in the evolutionary algorithm.}
 \item{"data"}{ - A data frame (tibble) containing the actual data (designs) of each Goelz Triangle in each generation.}
}
}
\description{
Creates conformity-optimized Goelz Triangle experimental designs
}
\details{
While \code{\link{goelz}} creates Goelz Triangle designs that roughly follow the conformity criterion set forth
by Goelz (2001), this function optimizes designs for conformity using an evolutionary algorithm. Function parameters other
than \code{N} are all controls on the evolutionary algorithm. The \code{\link{ecr}} package is used for the evolutionary
algorithm.
}
\examples{
dat <- goelz_optim()
}
\references{
\itemize{
 \item Goelz JCG (2001) Systematic experimental designs for mixed species plantings. Native Plants Journal 2:90–96.
 \url{http://npj.uwpress.org/content/2/2/90.short}
}
}
\seealso{
Other definition functions: \code{\link{goelz_add_border}},
  \code{\link{goelz_corners}}, \code{\link{goelz_guides}},
  \code{\link{goelz_mirror}}, \code{\link{goelz_starts}},
  \code{\link{goelz}}, \code{\link{nelder_biculture}},
  \code{\link{nelder_decision}}, \code{\link{nelder}}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
