% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{goelz_starts}
\alias{goelz_starts}
\title{Get coordinates of row starts for Goelz Triangle experimental design}
\usage{
goelz_starts(data)
}
\arguments{
\item{data}{An object of class goelz.}
}
\value{
A data frame (tibble).
}
\description{
Gets coordinates of row starts for Goelz Triangle experimental design.
}
\examples{
dat <- goelz()
goelz_starts(data = dat)
}
\seealso{
Other definition functions: \code{\link{goelz_add_border}},
  \code{\link{goelz_corners}}, \code{\link{goelz_guides}},
  \code{\link{goelz_mirror}}, \code{\link{goelz_optim}},
  \code{\link{goelz}},
  \code{\link{nelder_biculture_competition}},
  \code{\link{nelder_biculture}},
  \code{\link{nelder_decision}}, \code{\link{nelder}},
  \code{\link{select_optimal_goelz}}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
