% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mutInteger}
\alias{mutInteger}
\title{Integer mutation}
\usage{
mutInteger(ind, p, lower, upper)
}
\description{
Performs integer mutation between a lower and upper bound. Used by \code{\link{goelz_optim}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
