% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mutShuffle}
\alias{mutShuffle}
\title{Shuffle mutation}
\usage{
mutShuffle(ind, p)
}
\description{
Performs mutation by shuffling elements within a vector. Used by \code{\link{nelder_biculture_optim}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
