% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nelder_biculture}
\alias{nelder_biculture}
\title{Create a Biculture Nelder Fan experimental design}
\usage{
nelder_biculture(design, comps = NULL)
}
\arguments{
\item{design}{An object of class nelder created by \code{\link{nelder}}.}

\item{comps}{An option numeric vector containing the ratios of one species in each spoke.
This can effectively be used to create non-standard bi-culture designs that deviate from the Goelz (2001) approach.
If \code{NULL}, the default, then a linear
gradient of speies composition is used between monoculture extremes for each species.}
}
\value{
An object of classes sysd, nelder, and nelder-biculture. This is a list of 5 elements,
the first 3 of which are the same as for \code{\link{nelder}}, and the last 2 of which are:
\itemize{
 \item{"species.counts"}{ - An abject of class "table" containing the total counts of each species in the design.}
 \item{"spoke.composition"}{ - A data frame (tibble) containing the ratio and counts of each species by spoke.}
}
}
\description{
Creates a Biculture Nelder Fan experimental design.
}
\details{
The Nelder Fan or Nelder Wheel Type Ia experimental design (Nelder 1962) is an experimental design that
systematically varies plant desnity within a single plot, where the growing area around each plant has a constant
shape throughout the design but increases as radius increases. Goelz (2001) adapted this design to simultaneously
study the effect of species composition by superimposing a species gradient along the arc (Figure 6 of Goelz, 2001).
This function takes a Nelder Fan design from \code{\link{nelder}} and adds species identities to create
the Goelz (2001) biculture version. For Nelder Fan designs with \code{max.angle = 360}, species monoculture are
set to opposite poles of the circle, and the composition gradient occurs in two direction along either side of the
circle in between. For Nelder Fan designs with \code{max angle < 360}, the composition gradient occurs in one
direction between monoculture extremes at either edge of the design.
}
\examples{
nelder.design <- nelder(DN = 1000, D1 = 3000, N = 5)
dat <- nelder_biculture(design = nelder.design)
}
\references{
\itemize{
 \item Nelder JA (1962) New kinds of systematic designs for spacing experiments.
 Biometrics 18:283-307.
 \url{http://www.jstor.org/stable/2527473}
 \item Goelz J (2001) Systematic experimental designs for mixed species plantings.
 Native Plants Journal 2:90–96.
 \url{http://npj.uwpress.org/content/2/2/90.short}
}
}
\seealso{
Other definition functions: \code{\link{goelz_add_border}},
  \code{\link{goelz_corners}}, \code{\link{goelz_guides}},
  \code{\link{goelz_mirror}}, \code{\link{goelz_optim}},
  \code{\link{goelz_starts}}, \code{\link{goelz}},
  \code{\link{nelder_biculture_competition}},
  \code{\link{nelder_decision}}, \code{\link{nelder}}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
