% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nelder_biculture_class_check}
\alias{nelder_biculture_class_check}
\title{Check if an object if of class nelder-biculture}
\usage{
nelder_biculture_class_check(data)
}
\value{
An error.
}
\description{
Checks if an object if of class nelder-biculture. Used by many nelder functions.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
