% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nelder_biculture_competition}
\alias{nelder_biculture_competition}
\title{Calculate competition indices for a Nelder Fan experimental design}
\usage{
nelder_biculture_competition(design, search.radius = NULL)
}
\arguments{
\item{design}{An object of class nelder.}

\item{search.radius}{Search radius to use for calculating inverse distance weighted competition (m). If \code{NULL},
the default, then inverse distance weighted competition is not calculated.}
}
\value{
A data frame (tibble), which is the \code{plants} element of the nelder object passed to \code{design},
but with 4 additioanl columns:
\itemize{
 \item{"A.inv.dist"}{ - The inverse distance weighted competition felt by that individual due to nearby individuals
 of species A}
 \item{"B.inv.dist"}{ - The inverse distance weighted competition felt by that individual due to nearby individuals
 of species B}
 \item{"A.neighbors"}{ - The number of directly adjacent individuals of species A}
 \item{"B.neighbors"}{ - The number of directly adjacent individuals of species B}
}
}
\description{
Calculates competition indices for a Nelder Fan experimental design.
}
\examples{
dat <- nelder(DN         = 1000,
              D1         = 3000,
              N          = 5,
              tau        = 1,
              even       = TRUE,
              max.angle  = 360)
dat.bi <- nelder_biculture(dat)
nelder_biculture_competition(dat.bi)
}
\seealso{
Other definition functions: \code{\link{goelz_add_border}},
  \code{\link{goelz_corners}}, \code{\link{goelz_guides}},
  \code{\link{goelz_mirror}}, \code{\link{goelz_optim}},
  \code{\link{goelz_starts}}, \code{\link{goelz}},
  \code{\link{nelder_biculture}},
  \code{\link{nelder_decision}}, \code{\link{nelder}}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
