% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nelder_biculture_optim}
\alias{nelder_biculture_optim}
\title{Create a composition-density-conformity-optimized Biculture Nelder Fan experimental design}
\usage{
nelder_biculture_optim(
  data,
  save.path = ".",
  init.pop = NULL,
  MU = 100,
  LAMBDA = MU,
  MAX.GEN = 150,
  P.RECOMB = 1,
  RECOMB = 0.1,
  P.MUT = 1,
  MUT = 0.05
)
}
\arguments{
\item{data}{An object of class nelder-biculture.}

\item{save.path}{A character string of the path for where to save data after each generation.}

\item{init.pop}{If \code{NULL}, the default, then the initial population will be generated via \code{MU} calls
to \code{\link{nelder}}. Otherwise, a character string of the path to the file containing the data from which to
generate the initial population. This file should be the "data" output file that was saved by this function on a
previous call. The iniital population will be taken as the final population in this file.}

\item{MU}{The population size.}

\item{LAMBDA}{The number of offspring to produce in each generation.}

\item{MAX.GEN}{The number of generations to run the evolutionary algorithm.}

\item{P.RECOMB}{The probability of two parents to perform crossover.}

\item{RECOMB}{The crossover probability for each gene.}

\item{P.MUT}{The probability to apply mutation to a child.}

\item{MUT}{The mutation probability for each gene.}
}
\value{
An list containing:
\itemize{
 \item{"layout"}{ - A data frame (tibble) containing the base nelder design.}
 \item{"stats"}{ - A data frame (tibble) containing statistics on each generation in the evolutionary algorithm.}
 \item{"data"}{ - A data frame (tibble) containing the actual data (designs) of each Biculture Nelder Fan in eac generation.}
}
}
\description{
Creates a competition-conformity-optimized Biculture Nelder Fan experimental design.
}
\details{
While \code{\link{nelder_biculture}} creates Biculture Nelder Fan designs that follow a composition-conformity
criterion across spokes, this function creates designs that are optimized to also conform to a composition-density-conformity
criterion for species composition environments across arcs using an evolutionary algorithm. Function parameters other than
\code{data}, \code{save.path}, and, \code{init.pop} are all controls on the evolutionary algorithm.
The \code{\link{ecr}} package is used for the evolutionary algorithm.
In addition to the data returned in the environment, the "stats" and "data" data frames are written as csv files
to \code{save.path}.
}
\examples{
dat <- nelder()
dat.bi <- nelder_biculture(data = dat)
dat.bi.optim <- nelder_biculture_optim(data = dat.bi)
}
\seealso{
Other definition functions: 
\code{\link{goelz_add_border}()},
\code{\link{goelz_corners}()},
\code{\link{goelz_guides}()},
\code{\link{goelz_mirror}()},
\code{\link{goelz_optim}()},
\code{\link{goelz_starts}()},
\code{\link{goelz}()},
\code{\link{nelder_biculture_competition}()},
\code{\link{nelder_biculture}()},
\code{\link{nelder_decision}()},
\code{\link{nelder_interspoke_distance}()},
\code{\link{nelder}()},
\code{\link{select_optimal_goelz}()},
\code{\link{select_optimal_nelder_biculture}()}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
