% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nelder_calc}
\alias{nelder_calc}
\title{Calculate Nelder Fan design}
\usage{
nelder_calc(
  alpha,
  theta,
  tau,
  D1,
  N,
  max.angle,
  n.spokes,
  arc.borders,
  spoke.borders
)
}
\arguments{
\item{alpha}{Rate of change along the spokes.}

\item{theta}{The angle between the spokes.}

\item{tau}{The "rectangularity" proportion}

\item{D1}{Plant density within the first experimental arc (plants ha-1)
(i.e. upper extreme of experimental plant density range).}

\item{N}{Number of experimental arcs (i.e. number of densities to be tested within D1 to DN).}

\item{max.angle}{The maximum rotation (in degrees) of the design. If 360, then a full circle design will be created.}

\item{n.spokes}{The number of spokes.}

\item{arc.borders}{Number of border arcs on either extreme.}

\item{spoke.borders}{Number of border spokes on either extreme (only used if max.angle < 360).}
}
\value{
A list containing the plant and plot data.
}
\description{
Calculates Nelder Fan design.
Used within \code{\link{nelder}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
