% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nelder_class_check}
\alias{nelder_class_check}
\title{Check if an object if of class nelder}
\usage{
nelder_class_check(data)
}
\value{
An error.
}
\description{
Checks if an object if of class nelder. Used by many nelder definition functions.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
