% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nelder_fitness}
\alias{nelder_fitness}
\title{Calculate the fitness of a Nelder Fan biculture design}
\usage{
nelder_fitness(design, layout)
}
\value{
A numeric value.
}
\description{
Calculates the fitness of a Nelder Fan biculture design. Fitness is calculated to reflect the
uniformity of compeition environments present across experimental arcs.
Used within \code{\link{nelder_biculture}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
