% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nelder_interspoke_distance}
\alias{nelder_interspoke_distance}
\title{Calculate distance between two individuals in the same arc and adjacent spokes of a Nelder Fan experimental design}
\usage{
nelder_interspoke_distance(data, arc = data$plot$arcs)
}
\arguments{
\item{data}{An object of class nelder.}

\item{arc}{An numeric value indicating which arc of the design to calculate distance for.}
}
\value{
A numeric value of the distance (m).
}
\description{
Calculates distance between two individuals in the same arc and adjacent spokes of a Nelder Fan experimental design.
}
\examples{
dat <- nelder(DN         = 1000,
              D1         = 3000,
              N          = 5,
              tau        = 1,
              even       = TRUE,
              max.angle  = 360)
nelder_interspoke_distance(dat)
}
\seealso{
Other definition functions: 
\code{\link{goelz_add_border}()},
\code{\link{goelz_corners}()},
\code{\link{goelz_guides}()},
\code{\link{goelz_mirror}()},
\code{\link{goelz_optim}()},
\code{\link{goelz_starts}()},
\code{\link{goelz}()},
\code{\link{nelder_biculture_competition}()},
\code{\link{nelder_biculture_optim}()},
\code{\link{nelder_biculture}()},
\code{\link{nelder_decision}()},
\code{\link{nelder}()},
\code{\link{select_optimal_goelz}()},
\code{\link{select_optimal_nelder_biculture}()}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
