% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fitness_trajectory}
\alias{plot_fitness_trajectory}
\title{Plot a diagnositc display of the fitness trajectory}
\usage{
plot_fitness_trajectory(data, plot = TRUE)
}
\arguments{
\item{data}{An object of class 'goelz-optim' created via \code{\link{goelz_optim}} or
'nelder-optim' created via \code{\link{nelder_biculture_optim}}.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned.
If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
If \code{plot = TRUE}, returns a ggplot object, otherwise the data that would create the plot is returned.
}
\description{
Plots a diagnositc display of the fitness trajectory from the genetic algorithm run by
\code{\link{goelz_optim}} or \code{\link{nelder_biculture_optim}}.
}
\examples{
dat <- nelder()
dat.bi <- nelder_biculture(data = dat)
dat.bi.optim <- nelder_biculture_optimr(data = dat.bi)
plot_fitness_trajectory(data = dat.bi.optim)
}
\seealso{
Other plot functions: 
\code{\link{plot_goelz}()},
\code{\link{plot_nelder_biculture_competition}()},
\code{\link{plot_nelder}()}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{plot functions}
