% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_goelz}
\alias{plot_goelz}
\title{Plot a Goelz Triangle experimental design}
\usage{
plot_goelz(
  data,
  size = 4,
  fill = "species",
  color = "border",
  label = "none",
  fill.discrete.palette = c("grey50", "white", "#E69F00", "#56B4E9", "#009E73",
    "#F0E442", "#0072B2", "#D55E00", "#CC79A7"),
  color.discrete.palette = c("black", "green", "red", "blue", "orange", "purple"),
  fill.continuous.palette = ggplot2::scale_fill_viridis_c(option = "magma"),
  color.continuous.palette = ggplot2::scale_color_viridis_c(option = "magma"),
  corners = FALSE,
  guides = FALSE,
  legend = TRUE
)
}
\arguments{
\item{data}{An object of class 'goelz' created via \code{\link{goelz}} with \code{split} = FALSE.}

\item{size}{A numeric value indicating point size.}

\item{fill}{One of "species", "zone", or "border", indicating the column of \code{data} that specifies the point fill color.}

\item{color}{One of "species", "zone", or "border", indicating the column of \code{data} that specifies the point border color.
Use "none" for no variation in point borders.}

\item{label}{A character string indicating the column name of \code{data} to use for labels on each point.
Use "none" for no labels.}

\item{fill.discrete.palette}{A character vector indicating the color palette of used for point fill.}

\item{color.discrete.palette}{A character vector indicating the color palette of used for point borders.}

\item{fill.continuous.palette}{ggplot2 function for a continuous color palette used for point fill.}

\item{color.continuous.palette}{ggplot2 function for a continuous color palette used for point color}

\item{corners}{Logical indicating whether or not to plot red X's at each corner of the rectangle containing the design.}

\item{guides}{Logical indicating whether or not to plot red +'s as guides at the extremes of each row in the design.}

\item{legend}{Logical indicating whether or not to include a legend for fill and color values.}
}
\value{
A ggplot2 object.
}
\description{
Plots a Goelz Triangle experimental design.
}
\examples{
dat <- goelz(N = 35, reps = 1, split = FALSE)
plot_goelz(data = dat)
}
\seealso{
Other plot functions: 
\code{\link{plot_fitness_trajectory}()},
\code{\link{plot_nelder_biculture_competition}()},
\code{\link{plot_nelder}()}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{plot functions}
