% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nelder}
\alias{plot_nelder}
\title{Plot a Nelder Fan experimental design}
\usage{
plot_nelder(data, size = 3, fill.palette = c("black", "white"),
  species = TRUE, legend = FALSE, ex.area = FALSE, caption = FALSE)
}
\arguments{
\item{data}{An object of class 'nelder' created via \code{\link{nelder}}.}

\item{size}{A numeric value indicating point size.}

\item{fill.palette}{A character vector of length 2 indicating the fill colors of different species.}

\item{species}{Logical indicating whether or not to plot species as differnet colors
(see \code{\link{nelder_biculture}}).}

\item{legend}{Logical indicating whether or not to include a legend for fill colors.}

\item{ex.area}{Logical indicating whether or not and example growing area of one plant should be shown.}

\item{caption}{Logical indicating whether or not to include a caption with generic data about the design.}
}
\value{
A ggplot2 object.
}
\description{
Plots a Nelder Fan experimental design.
}
\examples{
dat <- nelder(DN         = 1000,
              D1         = 3000,
              N          = 5,
              tau        = 1,
              even       = TRUE,
              max.angle  = 360)
plot_nelder(data = dat)
}
\seealso{
Other plot functions: \code{\link{plot_goelz}}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{plot functions}
