% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{remove_edge}
\alias{remove_edge}
\title{Remove edge}
\usage{
remove_edge(data, side, n)
}
\value{
An object of class goelz.
}
\description{
Removes edge.
Used by \code{\link{goelz_mirror}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
