% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{run_ga}
\alias{run_ga}
\title{Run genetic algorithm}
\usage{
run_ga(
  control,
  population,
  layout,
  layout.comp,
  LAMBDA,
  MAX.GEN,
  P.RECOMB,
  RECOMB,
  P.MUT,
  save.path,
  start.gen
)
}
\value{
A list containing fitness stats and population data.
}
\description{
Runs genetic algorithm for \code{\link{goelz_optim}} & \code{\link{nelder_biculture_optim}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
