% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{select_optimal}
\alias{select_optimal}
\title{Select optimal individual}
\usage{
select_optimal(x)
}
\value{
An numeric vector of the optimal species composition.
}
\description{
Selects the optimal individual from an object of class goelz-optim or nelder-optim.
Used by \code{\link{select_optimal_goelz}} & \code{\link{select_optimal_nelder_biculture}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
