% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goelz.R
\name{select_optimal_goelz}
\alias{select_optimal_goelz}
\title{Select optimal Goelz Triangle design from the results of goelz_optim}
\usage{
select_optimal_goelz(optim.results)
}
\arguments{
\item{optim.results}{An object of class goelz-optim.}
}
\value{
A data frame (tibble) of class goelz.
}
\description{
Selects optimal Goelz Triangle design from the results of \code{\link{goelz_optim}}.
}
\examples{
optim.dat <- goelz_optim()
my_goelz <- select_optimal_goelz(optim.results = optim.dat)
}
\seealso{
Other definition functions: \code{\link{goelz_add_border}},
  \code{\link{goelz_corners}}, \code{\link{goelz_guides}},
  \code{\link{goelz_mirror}}, \code{\link{goelz_optim}},
  \code{\link{goelz_starts}}, \code{\link{goelz}},
  \code{\link{nelder_biculture_competition}},
  \code{\link{nelder_biculture}},
  \code{\link{nelder_decision}}, \code{\link{nelder}}
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\concept{definition functions}
