% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{yard_to_feet_frac}
\alias{yard_to_feet_frac}
\title{Convert yards to feet with fractional inches}
\usage{
yard_to_feet_frac(y)
}
\arguments{
\item{y}{A numeric value of length 1.}
}
\value{
A character string
}
\description{
Converts yards to feet with fractional inches.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
