% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{compile_pop}
\alias{compile_pop}
\title{Compile popluation data}
\usage{
compile_pop(pop, fit, layout, GEN)
}
\value{
A list.
}
\description{
Compiles popluation data.
Used by \code{\link{goelz_optim}} & \code{\link{nelder_biculture_optim}}.
}
\author{
Kevin J Wolz, \email{kevin@savannainstitute.org}
}
\keyword{internal}
